Modified the data in the Raise Your Rate table so that the balance at the point each rate raise happens is included.
Even though it's duplicate data -- the balance at the time of the rate raise is the same as the previous rate --
it more closely matches what goes on the chart. And it probably makes more sense.

Semantically, what would probably make the most sense is to use rowspan so the otherwise blank columns become part
of the columns above them. But that makes the programming job even harder. It would be best to first investigate
how rowspan affects accessibility.