On second thought, I think layers should be constructors, not objects.
Since a constructor is a function, existing layers won't need to change, and layers that
only draw things and aren't needed again can remain plain old functions.

A layer can return an object with functions like redraw(), but that's not an absolute
requirement.

Also, making the layer a constructor means the returned object is an individual object
that can have its own state.

This reverts commit eb568305908a8a124d4970d10bc57b743f7c22e8.